+++ cdk/kittyhawk/src/application/redis.ts
@@ -1,6 +1,6 @@
 import { Construct } from 'constructs';
-import { Application } from '.';
 import { DeploymentProps } from '../deployment';
+import { Application } from './base';
 
 export interface RedisApplicationProps {
   /**
@@ -12,7 +12,7 @@ export interface RedisApplicationProps {
    * Port to expose the application on.
    */
   readonly port?: number;
-  
+
   /**
    * Creates a service account and attach it to any deployment pods.
    * serviceAccountName: release name
@@ -20,7 +20,7 @@ export interface RedisApplicationProps {
   readonly createServiceAccount?: boolean;
 }
 
-export default class RedisApplication extends Application {
+export class RedisApplication extends Application {
   constructor(scope: Construct, appname: string, redisProps: RedisApplicationProps) {
     super(scope, appname, {
       deployment: {
